requirements.txt
torch==1.1.0
numpy
tqdm
matplotlib
visdom
pyyaml
tabulate